+++ example/android/app/src/main/kotlin/com/solusibejo/flutter_avo_inspector/example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.solusibejo.flutter_avo_inspector.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()