+++ example/android/app/src/main/kotlin/com/solusibejo/flutter_avo_inspector_example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.solusibejo.flutter_avo_inspector_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}